% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot the location of each incident in a given state and year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An integer for the code number of a state}

\item{year}{An integer or element coercible to integer representing a year.}
}
\value{
This function draws a map of the requested state, and plot on it the
incidents occured in that state in the requested year. Each point represents an incident
}
\description{
This function takes input of a state number and a year, and
plot the incidents of the given year in the map of the given state.
Conditions that may result an error:
\code{state.num} is not a valid number of states;
\code{year} is not a valid  number of year;
Data of requested year does not exist;
}
\examples{
\dontrun{fars_map_state(1, 2014)}

}
