% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Batch read data "MONTH" info in given year(s)}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a vector of integer representing number of year(s)}
}
\value{
This function returns a list of tibbles.
Each tibble contain two columns with month and year information extrated from data of requested years.
}
\description{
This function creates a list of tibbles based on input of years.
It will first read in the data corresponding to the year(s),
then extract the year and month information for each requested year.
Conditions that may result in an error:
\code{years} is not a valid vector containing number of year(s);
Data of requested year does not exist.
}
\examples{
\dontrun{fars_read_years(c(2013, 2014))}

}
