% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize number of observations in each month in given year(s)}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a vector of integer representing number of year(s)}
}
\value{
This function returns a list of tibbles, each tibble contains two columns.
The first column is each month, the second column is observation number in that month.
}
\description{
This function counts the number of observations in each month based on year input.
Conditions that may result in an error:
\code{years} is not a valid vector containing number of year(s);
Data of requested year does not exist.
}
\examples{
\dontrun{fars_summarize_years(c(2013, 2014))}

}
