% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Print a file name based on year input}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer or element coercible to integer representing a year.}
}
\value{
This function prints a character string of file name based on input year.
It can be used for reading the file.
}
\description{
Conditions that may result in an error:
\code{year} is not a valid number of a year;
Data of requested year does not exist.
}
\examples{
\dontrun{make_filename(2014)}

}
