% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/far_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state function}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{the number of the state for which we want
to draw the accident's coordinates.}

\item{year}{an integer indicating the year for which we want to draw the accident?s coordinates}
}
\value{
a plot with all accident's coordinates drawed of the /code{state.num}
and /code{year} or null if the /code{state.num} does not exist
}
\description{
given an state and a year plots the coordenates of the accidents in
this state an year
}
\examples{
fars_map_state(1,2013)

}
