% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{index of a state}

\item{year}{year to specify}
}
\value{
map plot of the number of accident. If there is no accidents for the year in the stat, return
"no accidents to plot".
}
\description{
Extract a data for a specific year and state.
}
\examples{
fars_map_state(state.num, 2015)

library(mapdata)
fars_map_state(12, 2014)
fars_map_state(36, 2014)

\dontrun{
fars_map_state(3, 2014)   # error
}


}
\references{
http://www.nhtsa.gov/Data/Fatality-Analysis-Reporting-System-(FARS)
}
