% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Draw a map and show the accidents on certain state and time}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An integral value}

\item{year}{A numeric value}
}
\value{
This function returns map to show the accidents' location.
}
\description{
This function shows the accidents' location whose LONGITUD value is not bigger than 900 and LATITUDE
value is not bigger than 90 on the map according to the input state number and year.
}
\details{
If the filename does not exist, it will return an error to tell you the filename does not exist.
         If the state number are not in the data, it will return an error to tell you the number are invalid.
         If no accident exist, it will give you a message that "no accidents to plot".
}
\examples{


}
