% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read.R
\name{fars_read}
\alias{fars_read}
\title{fars_read
This function accepts the location of the CSV file and reads it
If such a file does not exit then it will produce an error with the file
does not exit error.}
\usage{
fars_read(filename)
}
\arguments{
\item{a}{csv file that may or not exit in the given location.}
}
\value{
returns a dataframe if the csv file exits if not returns a file doesnot exit
error
}
\description{
fars_read
This function accepts the location of the CSV file and reads it
If such a file does not exit then it will produce an error with the file
does not exit error.
}
\examples{
\dontrun{accident <- fars_read('accident_2015.csv')}

}
\seealso{
\code{\link{make_filename}}
}
