% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/far_functions.R
\name{fars_read}
\alias{fars_read}
\title{fars_read function
read a csv file given a route to the file.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{an string indicating the complete route or relative route (from the working directory) to the file. The file can be compressed.}
}
\value{
a table_df that contents the data of csv. If the file doesn?t exist returns an error.
}
\description{
fars_read function
read a csv file given a route to the file.
}
\examples{
fars_read (accident_2013.csv.bz2)
fars_read (accident_2014.csv.bz2)

}
