% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{fars_read}
\usage{
fars_read(filename)
}
\arguments{
\item{file}{A character string giving the text the function will print}
}
\value{
Data object from the file specified by filename.
If file does not exist, return a message "file does not exists".
}
\description{
This is a simple function to read a file.
}
\examples{
fars_read(filename)

}
\references{
http://www.nhtsa.gov/Data/Fatality-Analysis-Reporting-System-(FARS)
}
