% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read a file into a data frame and then convert to a data frame tbl}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{a file path}
}
\value{
This function returns a data frame tbl.
}
\description{
This is a simple function that use the function \code{read_csv()} to read a file into a data frame
and then convert it to a data frame tbl.
}
\details{
If the filename does not exist, it will return an error to tell you the filename does not exist.
}
\examples{

}
