% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read_years.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years
Uses both make_filename(year) and fars_read(file)
Accepts one or more years as a list and produce the file name in the format of
accident_year.csv.bz2 and produce the relevent dataset}
\usage{
fars_read_years(years)
}
\arguments{
\item{one}{or more years as a list}
}
\value{
returns one or more datasets according to the number of years passed
returns null if the input year is invalid
}
\description{
fars_read_years
Uses both make_filename(year) and fars_read(file)
Accepts one or more years as a list and produce the file name in the format of
accident_year.csv.bz2 and produce the relevent dataset
}
\examples{
\dontrun{fars_read_years(2015)}
\dontrun{fars_read_years(1999:2015)}

}
