% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/far_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years fuction
given a list of years,  extracts for each  year the MONTH column
 and a year column of the accident data of that year.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a vector or list of integers representing the years}
}
\value{
a list of dataframes of the same length of /code {"year"}.
Each element of the list refers to a year of the list and si a dataframe that
includes the MONTH column of the accident data of the year and a column with the year.
If there is no data of the year the element of the list will be null.
}
\description{
fars_read_years fuction
given a list of years,  extracts for each  year the MONTH column
 and a year column of the accident data of that year.
}
\examples{
fars_read_years(c(2013, 2014))
 @importFrom dplyr mutate, select
 @export
}
