% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{Spcify years to read specific files.}
}
\value{
If years are not in valid range, "invalid year" will be returned.
}
\description{
Reading files for a single or multiple years.
}
\examples{
fars_read_years(2012)
fars_read_years(2012:2015)

}
\references{
http://www.nhtsa.gov/Data/Fatality-Analysis-Reporting-System-(FARS)
}
