% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read a list of the accident data which contains month and year}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a vector (atomic or list) of years}
}
\value{
This function returns a list of accident data of different years
}
\description{
This is a a simple function to read some years' accident data and then select the year and month of the
accidents.
}
\details{
If the filename does not exist, it will return an error to tell you the filename does not exist.
         If the input is not numeric, it will give a warning and then return NULL
}
\examples{

}
