% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_summarize_years.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years
Accepts one or more years as a list and count the numver if observations
grouping them by minth and year
Uses fars_read_years(years)}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{or more years as a list}
}
\value{
a dataframe of counts by months and years
}
\description{
fars_summarize_years
Accepts one or more years as a list and count the numver if observations
grouping them by minth and year
Uses fars_read_years(years)
}
\examples{
\dontrun{data_summary <- fars_summarize_years(2015)}

}
