% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/far_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years function
given a list of years gives the number of accidents by month of each of the years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a vector or list of ints or strings  representing years}
}
\value{
a table which variables are the years on list and an unique
file that indicates the number of accidents by month of this year
}
\description{
fars_summarize_years function
given a list of years gives the number of accidents by month of each of the years
}
\examples{
fars_summarize_years(c(2013, 2014, 2015))

}
