% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{years to summarize.}
}
\value{
Summary firls for specified years aggregated by months.
}
\description{
Summarizes files by years and month.
}
\examples{
fars_summarize_years(2014)
fars_summarize_years(2014:2015)

}
\references{
http://www.nhtsa.gov/Data/Fatality-Analysis-Reporting-System-(FARS)
}
