% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize the accident data according to accidents' year and month}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a vector (atomic or list) of years}
}
\value{
This function return a data frame tbl to show the number of accidents in the years of each month.
}
\description{
This function summarize the accident data and count the number of accidents in the years of each month.
}
\details{
If the filename does not exist, it will return an error to tell you the filename does not exist.
         If the input is not numeric, it will give a warning and then return NULL
}
\examples{

}
