% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_filename.R
\name{make_filename}
\alias{make_filename}
\title{make_filename
This function accepts a the year as a string value and returns file name
concatenating it as accident_year.csv.bz2}
\usage{
make_filename(year)
}
\arguments{
\item{the}{year as a string value}
}
\value{
returns the concatenated name with the year in the format
accident_year.csv.bz2
}
\description{
make_filename
This function accepts a the year as a string value and returns file name
concatenating it as accident_year.csv.bz2
}
\examples{
\dontrun{filename <- make_filename('2015')}

}
