% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/far_functions.R
\name{make_filename}
\alias{make_filename}
\title{make_filename function
 given a year, makes the complete name of the "accident" filename. It has the form "accident_year.csv.bz2"
 where year is substituted by /code{"year"}}
\usage{
make_filename(year)
}
\arguments{
\item{year}{an integer or string indicating the year to which we want to obtain the filename}
}
\value{
an string that represents the complete name of the filename.
}
\description{
make_filename function
 given a year, makes the complete name of the "accident" filename. It has the form "accident_year.csv.bz2"
 where year is substituted by /code{"year"}
}
\examples{
\dontrun{
 make_filename(2013)
 make_filename(2014)
}
}
