% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/far_functions.R
\name{make_filename}
\alias{make_filename}
\title{make_filename function
 given a year, makes the complete name of the "accident" filename. It has the form "accident_year.csv.bz2"
 where year is substituted by /code{"year"}
 @param year an integer or string indicating the year to which we want to obtain the filename
 @return  an string that represents the complete name of the filename.
 @examples
 make_filename(2013)   returns accident_2013.csv.bz2
 make_filename(2014)   returns accident_2014.csv.bz2}
\usage{
make_filename(year)
}
\description{
make_filename function
 given a year, makes the complete name of the "accident" filename. It has the form "accident_year.csv.bz2"
 where year is substituted by /code{"year"}
 @param year an integer or string indicating the year to which we want to obtain the filename
 @return  an string that represents the complete name of the filename.
 @examples
 make_filename(2013)   returns accident_2013.csv.bz2
 make_filename(2014)   returns accident_2014.csv.bz2
}
