% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{make_filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Specify years to read data}
}
\value{
Print created file name.
}
\description{
This is to create a file name as "accident_%d.csv.bz2" combined with a specific year.
}
\examples{
make_filename(2015)

}
\references{
http://www.nhtsa.gov/Data/Fatality-Analysis-Reporting-System-(FARS)
}
