% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Print a filename according to its year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A numeric value}
}
\value{
This function returns a phrase to print with the input year among it.
}
\description{
This is a simple function to print the filename of the accident data according to its year.
}
\examples{
make_filename(2013)

}
